Decomp scrcmd

+++ include/tv.h
@@ -1,12 +1,6 @@
 #ifndef GUARD_TV_H
 #define GUARD_TV_H
 
-// Exported type declarations
-
-// Exported RAM declarations
-
-// Exported ROM declarations
-
 extern u8 *const gTVStringVarPtrs[3];
 
 void ClearTVShowData(void);
@@ -16,6 +10,6 @@ bool8 GetPriceReduction(u8 newsKind);
 u8 GetRibbonCount(struct Pokemon *pokemon);
 void sub_80EE184(void);
 void sub_80F14F8(TVShow *shows);
-int sub_80EF370(int value);
+size_t sub_80EF370(int value);
 
 #endif //GUARD_TV_H
